+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,17 @@
 # sit-onyx
 
+## 0.1.0-alpha.6
+
+### Minor Changes
+
+- bd5040f: feat: add `OnyxIcon` component
+
+## 0.1.0-alpha.5
+
+### Minor Changes
+
+- 8d5c937: feat: add component `OnyxHeadline`
+
 ## 0.1.0-alpha.4
 
 ### Patch Changes