+++ packages/vitepress-theme/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sit-onyx/vitepress-theme
 
+## 1.0.0-alpha.5
+
+### Patch Changes
+
+- Updated dependencies [bd5040f]
+  - sit-onyx@0.1.0-alpha.6
+
+## 1.0.0-alpha.4
+
+### Patch Changes
+
+- Updated dependencies [8d5c937]
+  - sit-onyx@0.1.0-alpha.5
+
 ## 1.0.0-alpha.3
 
 ### Patch Changes